+++ src/main/java/com/solvd/carservice/domain/equipment/material/EngineOil.java
@@ -2,13 +2,11 @@ package com.solvd.carservice.domain.equipment.material;
 
 import com.solvd.carservice.domain.equipment.MaterialForRepair;
 
-import java.math.BigDecimal;
-
 public class EngineOil extends MaterialForRepair {
 
     private String engineType;
 
-    public  EngineOil(String name, BigDecimal quantity, String unit, BigDecimal priceOfUnit, String engineType) {
+    public  EngineOil(String name, Double quantity, String unit, Double priceOfUnit, String engineType) {
         super(name, quantity, unit, priceOfUnit);
         this.engineType = engineType;
     }
