+++ src/main/java/com/solvd/carservice/domain/equipment/material/ThermalProtection.java
@@ -2,13 +2,11 @@ package com.solvd.carservice.domain.equipment.material;
 
 import com.solvd.carservice.domain.equipment.MaterialForRepair;
 
-import java.math.BigDecimal;
-
 public class ThermalProtection extends MaterialForRepair {
 
     private Integer temperatureProtection;
 
-    public ThermalProtection (String name, BigDecimal quantity, String unit, BigDecimal priceOfUnit, Integer temperatureProtection) {
+    public ThermalProtection (String name, Double quantity, String unit, Double priceOfUnit, Integer temperatureProtection) {
         super(name, quantity, unit, priceOfUnit);
         this.temperatureProtection = temperatureProtection;
     }
